refactor(footer.tsx): move footercontext to seperate file
this pr refactor code inside the footer componenet and move the constant footercontext to seperate
file inside the constants folder i have created to add more readibility to the code
